requirements.txt
Flask
transformers
torch
python-dotenv